refactor: extract risk.rs and baseline.rs from brain/metrics.rs
Split metrics.rs (2,420 lines) into 3 focused modules:
- risk.rs (241 lines): RiskTier enum, classify_risk()
- baseline.rs (324 lines): rules_baseline_classify(), print_baseline()
- metrics.rs (1,885 lines): remaining metrics, re-exports both

All existing import paths preserved via pub use re-exports.
